financial_analysis/requirements.txt
# Financial Statement Analysis - PDF to Excel/Word
pdfplumber>=0.10.0
openpyxl>=3.1.0
python-docx>=1.0.0
pandas>=2.0.0
pytesseract>=0.3.10
pypdfium2>=4.30.0
